refactor: input-price.js -> input-dual-range.js
